make Images.tsx DRYer by extracting animations to ImageAnimation
